ft_puthexa

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: mle-biha <mle-biha@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/21 15:50:15 by mle-biha          #+#    #+#             */
-/*   Updated: 2022/08/17 21:19:25 by mle-biha         ###   ########.fr       */
+/*   Updated: 2022/08/18 10:40:06 by mle-biha         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@
 
 int		ft_printf(const char *s, ...);
 int		ft_putchar(int c);
-int		ft_puthexa(int nb, char *base);
+int		ft_puthexa(size_t nb, char *base);
 int		ft_putnbr(int n);
 int		ft_putstr(char *s);
 int		ft_putunsigned(unsigned int n);
